pyproject.toml
[project]
name = "aiwf-eval-tinkering"
version = "0.1.0"
description = "Eval runner for AI Engineer World's Fair conversations"
requires-python = ">=3.12"
dependencies = [
  "pipecat-ai[google,openai]==0.0.84",
  "python-dotenv",
  "loguru",
  "openai>=1.0.0",
  "google-genai>=0.2.0",
]
